+++ docs/task-flow.md
@@ -782,7 +782,7 @@ Tags (IN/OUT/INOUT/…) are used by `Orchestrator::submit_*` to derive TensorMap
 dependencies and nothing else. Scheduler, WorkerThread, child, runtime.so, and
 kernels do not inspect them. Keeping tags only in Layer ① simplifies the blob
 and makes the "tags are Orchestrator input" rule explicit. Matches existing
-runtime: `ChipStorageTaskArgs` (`task_args.h:157`) is already declared with
+runtime: `ChipStorageTaskArgs` (`task_args.h`) is already declared with
 `void` as the TensorTag parameter.
 
 ### Why no `WorkerPayload` wrapper
@@ -833,6 +833,8 @@ lives in the mailbox blob bytes on the child side — view doesn't care.
 - [chip-level-arch.md](chip-level-arch.md) — L2 single-chip: three-program
   model (host / AICPU / AICore)
 - [`../src/common/task_interface/task_args.h`](../src/common/task_interface/task_args.h)
-  — `TaskArgs` template and `ChipStorageTaskArgs` alias
+  — `TaskArgsTpl` template and the `ChipStorageTaskArgs` alias
+- [`../src/common/task_interface/task_args_wire.h`](../src/common/task_interface/task_args_wire.h)
+  — the L3+ `TaskArgs`, `TaskArgsView`, and the mailbox blob codec
 - [`../src/common/task_interface/tensor.h`](../src/common/task_interface/tensor.h)
   — `ChipTensor` POD and `TensorArgType` enum
