+++ src/common/platform/sim/host/c_api_shared.cpp
@@ -27,7 +27,7 @@
 #include "call_config.h"
 #include "device_runner_base.h"
 #include "prepare_callable_common.h"
-#include "task_args.h"
+#include "task_args_wire.h"
 #include "native_run_context.h"
 
 #include <dlfcn.h>
